requirements.txt
Django==3.2.9
psycopg2==2.9.1
black==21.10b0
isort==5.10.1
djangorestframework==3.12.4
coverage==6.1.1
pytest==6.2.5
pytest-django==4.4.0
factory-boy==3.2.1
djangorestframework-simplejwt==5.0.0
django-sendgrid-v5==1.1.1
